Create  Description
This is a job simulation of Accenture via the Forage platform where we step into the shoes of a data analyst. Here we clean the dataset, analyse the dataset and present the insights to the stakeholder. Our job is to analyse user behaviour and other such metrics for the social media app SocialBuzz.